recipe: Fix naming issue with `libicudata.*.*.dylib` on Mac.
Previously, the `libicudata.*.*.dylib` library on Mac would keep a
relative path to itself. In particular, it had a path to
`../lib/libicudata.*.*.dylib`, which was a bit strange and caused
linking errors for any binary or library that linked against it. As it
was a relative path, `conda-build` didn't pick up on the fact that this
should be changed to use `@rpath` as it would normally. Here we add the
flag `--enable-rpath`, which seems to force `libicudata.*.*.dylib` to
embed the full path to itself. This should correct that problem and end
the linking issues that every other package has experienced with this
library on Mac.